requirements.txt
glcontext==3.0.0
llvmlite==0.44.0
moderngl==5.12.0
numba==0.61.0
numpy==2.1.3
pillow==11.1.0
pygame-ce==2.5.2
PyGLM==2.7.3
